refactor(product-collection): Moved persistense directory into infrastructure
