+++ module/product-collection/src/Infrastructure/Persistence/Projector/ProductCollectionElement/DbalProductCollectionElementRemovedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ProductCollection\Persistence\Dbal\Projector\ProductCollectionElement;
+namespace Ergonode\ProductCollection\Infrastructure\Persistence\Projector\ProductCollectionElement;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -17,7 +17,7 @@ use Ergonode\ProductCollection\Domain\Event\ProductCollectionElementRemovedEvent
 
 /**
  */
-class ProductCollectionElementRemovedEventProjector
+class DbalProductCollectionElementRemovedEventProjector
 {
     private const TABLE_ELEMENT = 'product_collection_element';
     private const TABLE_COLLECTION = 'product_collection';
