+++ module/product-collection/src/Infrastructure/Persistence/Projector/ProductCollectionType/DbalProductCollectionTypeCreatedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ProductCollection\Persistence\Dbal\Projector\ProductCollectionType;
+namespace Ergonode\ProductCollection\Infrastructure\Persistence\Projector\ProductCollectionType;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -16,7 +16,7 @@ use JMS\Serializer\SerializerInterface;
 
 /**
  */
-class ProductCollectionTypeCreatedEventProjector
+class DbalProductCollectionTypeCreatedEventProjector
 {
     private const TABLE = 'product_collection_type';
