+++ module/product-collection/src/Infrastructure/Persistence/Query/DbalProductCollectionQuery.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ProductCollection\Persistence\Dbal\Query;
+namespace Ergonode\ProductCollection\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
